WFCORE-5493: Invoke DeploymentTransformer in "full-replace-deployment"
