+++ server/src/main/java/org/jboss/as/server/deployment/DeploymentAddHandler.java
@@ -39,8 +39,6 @@ import static org.jboss.as.server.deployment.DeploymentHandlerUtils.hasValidCont
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.Iterator;
-import java.util.ServiceLoader;
 
 import org.jboss.as.controller.AttributeDefinition;
 import org.jboss.as.controller.OperationContext;
@@ -53,8 +51,6 @@ import org.jboss.as.controller.descriptions.ModelDescriptionConstants;
 import org.jboss.as.controller.registry.Resource;
 import org.jboss.as.protocol.StreamUtils;
 import org.jboss.as.repository.ContentRepository;
-import org.jboss.as.server.controller.resources.DeploymentAttributes;
-import org.jboss.as.server.deployment.transformation.DeploymentTransformer;
 import org.jboss.as.server.logging.ServerLogger;
 import org.jboss.dmr.ModelNode;
 
@@ -71,22 +67,16 @@ public class DeploymentAddHandler implements OperationStepHandler {
 
     protected final ContentRepository contentRepository;
 
-    @SuppressWarnings("deprecation")
-    private final DeploymentTransformer deploymentTransformer;
+    private final DeploymentTransformationUtil deploymentTransformationUtil;
 
     protected DeploymentAddHandler(final ContentRepository contentRepository) {
         assert contentRepository != null : "Null contentRepository";
         this.contentRepository = contentRepository;
-        this.deploymentTransformer = loadDeploymentTransformer();
+        this.deploymentTransformationUtil = new DeploymentTransformationUtil();
     }
 
     public static DeploymentAddHandler create(final ContentRepository contentRepository) {
         return new DeploymentAddHandler(contentRepository);
-    }
-
-    private static DeploymentTransformer loadDeploymentTransformer() {
-        Iterator<DeploymentTransformer> iter = ServiceLoader.load(DeploymentTransformer.class, DeploymentAddHandler.class.getClassLoader()).iterator();
-        return iter.hasNext() ? iter.next() : null;
     }
 
     /**
@@ -215,7 +205,7 @@ public class DeploymentAddHandler implements OperationStepHandler {
         InputStream transformed = null;
         try {
             try {
-                transformed = deploymentTransformer == null ? in : doTransformation(context, contentItemNode, name, in);
+                transformed = deploymentTransformationUtil.doTransformation(context, contentItemNode, name, in);
                 hash = contentRepository.addContent(transformed);
             } catch (IOException e) {
                 throw createFailureException(e.toString());
@@ -226,25 +216,6 @@ public class DeploymentAddHandler implements OperationStepHandler {
             StreamUtils.safeClose(transformed);
         }
         return new DeploymentHandlerUtil.ContentItem(hash);
-    }
-
-    private InputStream doTransformation(OperationContext context, ModelNode contentItemNode, String name, InputStream in) throws IOException, OperationFailedException {
-        try {
-            return deploymentTransformer.transform(in, name);
-        } catch (RuntimeException t) {
-            // Check if the InputStream is already attached to the operation request (as per CONTENT_INPUT_STREAM_INDEX check) and ignore that case
-            // as calling getInputStream would of returned the already partially consumed InputStream.
-            // Also verify that the thrown exception is the specific WFCORE-5198 `Error code 3`.
-            if (!contentItemNode.hasDefined(DeploymentAttributes.CONTENT_INPUT_STREAM_INDEX.getName()) &&
-                    t.getCause() != null && t.getCause().getCause() != null &&
-                    t.getCause().getCause() instanceof IOException &&
-                    t.getCause().getCause().getMessage().contains("during transformation. Error code 3")) {
-                ServerLogger.ROOT_LOGGER.tracef(t, "Ignoring transformation error and using original archive %s", name);
-                return getInputStream(context, contentItemNode);
-            } else {
-                throw t;
-            }
-        }
     }
 
     DeploymentHandlerUtil.ContentItem addUnmanaged(ModelNode contentItemNode) throws OperationFailedException {
