+++ server/src/main/java/org/jboss/as/server/deployment/DeploymentFullReplaceHandler.java
@@ -65,9 +65,12 @@ public class DeploymentFullReplaceHandler implements OperationStepHandler {
 
     protected final ContentRepository contentRepository;
 
+    private final DeploymentTransformationUtil deploymentTransformationUtil;
+
     protected DeploymentFullReplaceHandler(final ContentRepository contentRepository) {
         assert contentRepository != null : "Null contentRepository";
         this.contentRepository = contentRepository;
+        this.deploymentTransformationUtil = new DeploymentTransformationUtil();
     }
 
     public static DeploymentFullReplaceHandler create(final ContentRepository contentRepository) {
@@ -122,7 +125,7 @@ public class DeploymentFullReplaceHandler implements OperationStepHandler {
             ContentReference reference = ModelContentReference.fromModelAddress(address, newHash);
             contentItem = addFromHash(reference);
         } else if (hasValidContentAdditionParameterDefined(contentItemNode)) {
-            contentItem = addFromContentAdditionParameter(context, contentItemNode);
+            contentItem = addFromContentAdditionParameter(context, contentItemNode, name);
             newHash = contentItem.getHash();
 
             // Replace the content data
@@ -196,18 +199,21 @@ public class DeploymentFullReplaceHandler implements OperationStepHandler {
         return new DeploymentHandlerUtil.ContentItem(reference.getHash());
     }
 
-    DeploymentHandlerUtil.ContentItem addFromContentAdditionParameter(OperationContext context, ModelNode contentItemNode) throws OperationFailedException {
+    DeploymentHandlerUtil.ContentItem addFromContentAdditionParameter(OperationContext context, ModelNode contentItemNode, String name) throws OperationFailedException {
         byte[] hash;
         InputStream in = getInputStream(context, contentItemNode);
+        InputStream transformed = null;
         try {
             try {
-                hash = contentRepository.addContent(in);
+                transformed = deploymentTransformationUtil.doTransformation(context, contentItemNode, name, in);
+                hash = contentRepository.addContent(transformed);
             } catch (IOException e) {
                 throw createFailureException(e.toString());
             }
 
         } finally {
             StreamUtils.safeClose(in);
+            StreamUtils.safeClose(transformed);
         }
         contentItemNode.clear(); // AS7-1029
         contentItemNode.get(CONTENT_HASH.getName()).set(hash);
